Create database initalizer

+++ backend/database/Models/ColorDb.cs
@@ -11,7 +11,7 @@ namespace database.Models {
     public class ColorDb {
         public UInt64 Id { get; set; }
         public String? Name { get; set; }
-        public UInt16 Value { get; set; }
+        public UInt32 Value { get; set; }
         public UInt64? UserId { get; set; }
         public UserDb? User { get; set; }
         public ICollection<WalletDb> Wallets { get; set; }
